Add Swift version
This does not quite match the screenshot in the repo. The problem I
think, unless I converted it wrong which is totally possible, is the
handle color referrs to the handle point color AND the line color
connecting the handle and main point.

the way to fix this would be to create an additional argument for
drawPathElement(_ Int, withPreviousPoint: NSPoint, inColor: NSColor,
withHandlesInColor: NSColor) that also specified the line color. Then
the point color of the handels could match the main point color and the
lines could be a separate color.